backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.8
fuzzywuzzy==0.18.0
h11==0.14.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
unicorn==2.1.2
urllib3==2.3.0
uvicorn==0.34.0